Abstract DQN model creation to separate it out
* Rename DeepQNetwork to DeepQModel as a model can contains networks, but not otherwise
* Generalise optimizer creation for loading model
* Make DeepQNetwork an abstract class in deep_q_networks.py
* Create a derived DQN class, which is the fully-connected deep-Q network used all along
* Add hyper-parameter to choose which DQN-template to use
